backend/CineBrain/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
numpy==2.3.1
packaging==25.0
pandas==2.3.1
psycopg==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
six==1.17.0
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.2
uritemplate==4.2.0
